box/Software/box-config-daemon/pify/requirements.txt
bottle>=0.12.13
python-networkmanager>=2.0
git+git://github.com/Robpol86/libnl.git
